requirements.txt
gradio 
fpdf
openai
python-docx
huggingface-hub==0.24.6 